Moved SortDirection enums to worker
