+++ src/components/Overlays/SortForm/index.tsx
@@ -5,7 +5,7 @@ import { useTranslations } from 'next-intl';
 import React, { useState, useEffect } from 'react';
 import Lightbox from '@/components/Lightbox';
 import { useSortForm } from '@/hooks/useSortForm';
-import { SortDirection } from '@/tools/sortby';
+import { SortDirection } from '@/workers/itemsIndexedDbWorker/schemas';
 
 interface Sortable {
   title: string;
