+++ src/hooks/useSortForm.ts
@@ -1,5 +1,5 @@
 import { useFiltersStore } from '@/stores/stores';
-import { type SortDirection } from '@/tools/sortby';
+import { type SortDirection } from '@/workers/itemsIndexedDbWorker/schemas';
 
 interface UseSortForm {
   visible: boolean;
